backend/requirements.txt
python-multipart~=0.0.6
fastapi~=0.104.0
uvicorn~=0.23.2
aiofiles~=23.2.1
Pillow~=10.1.0
psycopg2~=2.9.9
SQLAlchemy~=2.0.23
alembic~=1.13.0
jsonpickle~=3.0.2
bcrypt==4.0.1
passlib[bcrypt]
python-jose~=3.3.0
scikit-learn~=1.3.2
numpy~=1.26.2
cryptography~=41.0.7
pydantic-settings~=2.1.0